Integration tests

+++ tasks/plugin/plugin.sh
@@ -16,6 +16,8 @@ RESELLER=$4
 
 GITHUBTOKEN=$5
 
+TEST=$6
+
 #Variables that might be set but have empty values should be unset to simplify logic below
 if [[ $SHA1 ]]; then
 	if [[ ${SHA1} ]]; then
@@ -75,15 +77,17 @@ logger -s "STASHPOP=$STASHPOP"
 logger -s "SHA1=$SHA1"
 logger -s "RESELLER=$RESELLER"
 logger -s "GITHUBTOKEN=$GITHUBTOKEN"
-
+logger -s "TEST=$TEST"
 
 cd $PLUGINPATH
 
 logger -s "CURRENT SHA1=$(git rev-parse HEAD)"
 
 if [ -n $STASHPOP ]; then
-	logger -s "STASHING"
-	git stash
+	if [ -z ${TEST} ]; then
+		logger -s "STASHING"
+		git stash
+	fi
 fi
 
 #Make sure all remotes are up to date before we pull or checkout SHA's which might not exist
@@ -102,15 +106,18 @@ else
 	logger -s "No SHA1 was specified so checking out $BRANCH and pulling"
 	git checkout $BRANCH
 	git pull
+
 fi
 
 if [[ $STASHPOP ]]; then
 
-	git stash pop
+	if [ -z ${TEST} ]; then
+		git stash pop
+	fi
 
 fi
 
-
+#Save current SHA1 for CodistoConnect repo
 cd $PLUGINPATH
 SHA1=`git rev-parse HEAD`
 SDATE=`date +%Y-%m-%d:%H:%M:%S`
@@ -148,7 +155,10 @@ if [ $BRANCH = "master" ] && [ -z $RESELLER ];	then
 	mv "$PLUGINPATH/code/community/Codisto/Sync/data/codisto_setup/$DATAINSTALLFNAME" "$PLUGINPATH/code/community/Codisto/Sync/data/codisto_setup/data-install-$PLUGINVERSION.php"
 
 	logger -s "Commiting change to data-install"
-	git add "$PLUGINPATH/code/community/Codisto/Sync/data/codisto_setup/data-install-$PLUGINVERSION.php" -f
+
+	if [ -z ${TEST} ]; then
+		git add "$PLUGINPATH/code/community/Codisto/Sync/data/codisto_setup/data-install-$PLUGINVERSION.php" -f
+	fi
 
 	#Lets update mysql4-install in code/community/Codisto/Sync/sql/codisto_setup must have matching file suffix to the plugin version so let's rename it to match
 
@@ -161,43 +171,76 @@ if [ $BRANCH = "master" ] && [ -z $RESELLER ];	then
 	mv "$PLUGINPATH/code/community/Codisto/Sync/sql/codisto_setup/$MYSQLINSTALLFNAME" "$PLUGINPATH/code/community/Codisto/Sync/sql/codisto_setup/mysql4-install-$PLUGINVERSION.php"
 	logger -s "Commiting change to mysql4-install"
 
-	git add "$PLUGINPATH/code/community/Codisto/Sync/sql/codisto_setup/mysql4-install-$PLUGINVERSION.php" --force
+	if [ -z ${TEST} ]; then
+		git add "$PLUGINPATH/code/community/Codisto/Sync/sql/codisto_setup/mysql4-install-$PLUGINVERSION.php" --force
+	fi
 
 	#Generate a new CHANGELOG.md and add it
 	logger -s "Generating new changelog"
 	github_changelog_generator --token $GITHUBTOKEN
-	git add CHANGELOG.md --force
+
+	if [ -z ${TEST} ]; then
+		git add CHANGELOG.md --force
+	fi
 
 	#Update config.xml with new plugin version
 	sed -ri "/1/s/([0-9]+.[0-9]+.[0-9]+)/$PLUGINVERSION/" "$PLUGINPATH/code/community/Codisto/Sync/etc/config.xml"
-	logger -s "git add $PLUGINPATH/code/community/Codisto/Sync/etc/config.xml"
-	git add "$PLUGINPATH/code/community/Codisto/Sync/etc/config.xml"
 
-	logger -s "Committing version bumped files"
-	git commit -am "BOT - Update data-install.php , bump plugin version and generate new changelog"
+	if [ -z ${TEST} ]; then
 
-	#Update master
-	git branch -D master
-	git checkout -b master
-	git push origin master --force
+		logger -s "git add $PLUGINPATH/code/community/Codisto/Sync/etc/config.xml"
+		git add "$PLUGINPATH/code/community/Codisto/Sync/etc/config.xml"
 
-	#Create a new tag with the plugin version
-	git tag -d "$PLUGINVERSION"
-	git tag -a "$PLUGINVERSION" -m "Version $PLUGINVERSION"
-	git push origin "$PLUGINVERSION" --force
-	
-	#checkout development branch of the specific SHA
-	logger -s "Checking out development branch with sha of $DEVSHA"
-	git checkout $DEVSHA --force
-	git merge "$PLUGINVERSION" --no-edit
-	
-	#we are in a detached state, kill development and recreate
-	git branch -D development
-	git checkout -b development
-	git push origin development --force
+		logger -s "Committing version bumped files"
+		git commit -am "BOT - Update data-install.php , bump plugin version and generate new changelog"
 
-	#We are finished with development so back to the SHA1 that was checked out for master
-	git checkout master
+		#Update master
+		git branch -D master
+		git checkout -b master
+
+		logger -s "Pusing master"
+		git push origin master --force
+	else
+		logger -s "Not pushing master - integration tests running"
+
+	fi
+
+	if [ -z ${TEST} ]; then
+
+			#Create a new tag with the plugin version
+			git tag -d "$PLUGINVERSION"
+			git tag -a "$PLUGINVERSION" -m "Version $PLUGINVERSION"
+			logger -s "Pusing tag"
+			git push origin "$PLUGINVERSION" --force
+
+	else
+
+			logger -s "Not pushing tag - integration tests running"
+
+	fi
+
+  if [ -z ${TEST} ]; then
+
+			#checkout development branch of the specific SHA
+			logger -s "Checking out development branch with sha of $DEVSHA"
+			git checkout $DEVSHA --force
+			git merge "$PLUGINVERSION" --no-edit
+
+			#we are in a detached state, kill development and recreate
+			git branch -D development
+			git checkout -b development
+
+			logger -s "Pusing development"
+			git push origin development --force
+
+			#We are finished with development so back to the SHA1 that was checked out for master
+		  git checkout master
+	else
+
+		logger -s "Not pushing development - integration tests running"
+
+	fi
+
 	cd $PLUGINPATH
 
 fi
@@ -279,32 +322,39 @@ fi
 if [ -z $RESELLER ]; then
 	if  [ "$BRANCH" == "development" ]  || [ "$BRANCH" == "master" ]; then
 
-		#Make a symlink so location is always the same when browser GET's the plugin (only if dev or master)
-		logger -s "Making a symlink from $SCRIPTPATH/$PLUGINFNAME  to $SCRIPTPATH/CodistoConnect$SUFFIX.tgz"
-		rm "$SCRIPTPATH/CodistoConnect$SUFFIX.tgz" >/dev/null 2>&1; ln -s "$SCRIPTPATH/$PLUGINFNAME" "$SCRIPTPATH/CodistoConnect$SUFFIX.tgz"
+		if [ -z $TEST ]; then
+			#Make a symlink so location is always the same when browser GET's the plugin (only if dev or master)
+			logger -s "Making a symlink from $SCRIPTPATH/$PLUGINFNAME  to $SCRIPTPATH/CodistoConnect$SUFFIX.tgz"
+			rm "$SCRIPTPATH/CodistoConnect$SUFFIX.tgz" >/dev/null 2>&1; ln -s "$SCRIPTPATH/$PLUGINFNAME" "$SCRIPTPATH/CodistoConnect$SUFFIX.tgz"
 
-		logger -s "Copying to codisto.com"
-		scp "$SCRIPTPATH/$PLUGINFNAME" "nwo@aws-chaos-us-w-1:/home/nwo/codisto/plugin/"
+			logger -s "Copying to codisto.com"
+			scp "$SCRIPTPATH/$PLUGINFNAME" "nwo@aws-chaos-us-w-1:/home/nwo/codisto/plugin/"
 
-		#Create symlink to this new version on codisto.com
-		if [ "$BRANCH" == "master" ]; then
-			ssh nwo@aws-chaos-us-w-1 "rm /home/nwo/codisto/plugin/CodistoConnect.tgz >/dev/null 2>&1; ln -s /home/nwo/codisto/plugin/$PLUGINFNAME /home/nwo/codisto/plugin/CodistoConnect.tgz"
+			#Create symlink to this new version on codisto.com
+			if [ "$BRANCH" == "master" ]; then
+				ssh nwo@aws-chaos-us-w-1 "rm /home/nwo/codisto/plugin/CodistoConnect.tgz >/dev/null 2>&1; ln -s /home/nwo/codisto/plugin/$PLUGINFNAME /home/nwo/codisto/plugin/CodistoConnect.tgz"
+			fi
+
+			if [ "$BRANCH" == "development" ]; then
+				ssh nwo@aws-chaos-us-w-1 "rm /home/nwo/codisto/plugin/CodistoConnect-beta.tgz >/dev/null 2>&1; ln -s /home/nwo/codisto/plugin/$PLUGINFNAME /home/nwo/codisto/plugin/CodistoConnect-beta.tgz"
+			fi
 		fi
-
-		if [ "$BRANCH" == "development" ]; then
-			ssh nwo@aws-chaos-us-w-1 "rm /home/nwo/codisto/plugin/CodistoConnect-beta.tgz >/dev/null 2>&1; ln -s /home/nwo/codisto/plugin/$PLUGINFNAME /home/nwo/codisto/plugin/CodistoConnect-beta.tgz"
-		fi
-
 	fi
 fi
 
-
 logger -s "Plugin build completed"
 
+if [ -z ${TEST} ]; then
+	#Leave a note about build
+	echo "extension built on $SDATE (branch [$BRANCH] - sha1[$SHA1]) - Extension package is -> $PLUGINFNAME">> $SCRIPTPATH/extensionbuildlog.txt
+else
 
-#Leave a note about build
-echo "extension built on $SDATE (branch [$BRANCH] - sha1[$SHA1]) - Extension package is -> $PLUGINFNAME">> $SCRIPTPATH/extensionbuildlog.txt
+	logger -s "Restoring system state after test run"
+	#restore filesystem state as we were only testing (its OK as the plugin builder only has one child so only testing or legitimate building can happen at once)
+
+  logger -s "PLUGIN PATH IS $PLUGINPATH"
+	cd $PLUGINPATH && git reset --hard && git clean -dfx --force
+fi
+
 #Leave plugin version and path to plugin as last line in STDOUT to be captured
 echo "$PLUGINVERSION~~$PLUGINFNAME"
-
-
